testing new interrupts

+++ PIC24_Ultrasonic.X/frontUltrasonic.h
@@ -27,5 +27,5 @@ void initFrontUltras(void);
 void __attribute__((__interrupt__, auto_psv)) _IC1Interrupt(void);
 void __attribute__((__interrupt__, auto_psv)) _IC2Interrupt(void);
 
-long convertToDistance(long time);
+unsigned convertToDistance(unsigned time);
 unsigned findObject(void);
